transaction-service-changes-commit-by-vamshi-kiran
Transaction-Service:
This service is responsible for transferring funds from a customer’s wallet to a merchant’s account.
Before processing the transfer, it verifies the customer, account, and merchant details within the system using REST API calls through RestTemplate. Once all validations pass and the transaction is completed, it returns a success response indicating that the fund transfer was successful.